migrate functions and static var to header
